pyproject.toml
[project]
name = "content-hub-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.0.0",
    "python-telegram-bot>=20.7",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.4"
]
